OCR Application Online/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.3.0
opencv-python-headless>=4.9.0.80
pdf2image>=1.17.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
